// Trinket-And-Casa/api/src/types.ts
export type SignupType = {
  name: string;
  email: string;
  phone_no: string;
  password: string;
};

export interface IError extends Error {
  code?: number;
  message: string;
  statusCode: number;
  isOperational: boolean;
  isCustomError: boolean;
  key: any[];
}

export interface IGAuth {
  access_token: string;
  id_token: string;
}

export interface IGAuthProfile {
  id: string;
  email: string;
  name: string;
  picture: string;
}

export type Role = "Admin" | "User";

export interface ITokenInfo {
  id: number;
  role: Role;
}

export interface CustomRequest extends Request {
  token_info?: ITokenInfo;
}